fix(lint): split ui main.rs family under the 500-line gate
